requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
pypdf>=3.0.0
pdf2image>=1.16.0
pillow>=10.0.0
pytesseract>=0.3.10
nltk>=3.8.0
indic-nlp-library>=0.91
langdetect>=1.0.9
python-dotenv>=1.0.0
tqdm>=4.65.0
sacrebleu>=2.3.1
rouge-score>=0.1.2
indic-transliteration>=2.0.0
pymongo>=4.5.0
groq>=0.3.0
